requirements.txt
flask
mysql-connector-python
paho-mqtt